Criacao do Dockerfile já buildando a aplicacao e iniciando

+++ src/main/java/com/nuvem/computacao/Atividade/repository/ArquivoRepository.java
@@ -3,5 +3,5 @@ package com.nuvem.computacao.Atividade.repository;
 import com.nuvem.computacao.Atividade.models.Arquivo;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ArquivoRepository extends JpaRepository<Arquivo,Integer> {
+public interface ArquivoRepository extends JpaRepository<Arquivo,Long> {
 }
